feat: move kubernetes_cleanup_jobs's config to top level script
